+++ paddle/fluid/framework/tensor_util.cc
@@ -998,278 +998,7 @@ void TensorFromStream(std::istream& is,
   }
 }
 
-// get tensor data point by DLDataType
-void* GetDstPtrByDLDataType(DLDataType type,
-                            phi::DenseTensor* dst,
-                            const platform::Place& dst_place) {
-  // vector types not currently supported
-  PADDLE_ENFORCE_LE(type.lanes,
-                    1,
-                    platform::errors::Unimplemented(
-                        "Vector type is not supported currently."));
-
-  switch (type.bits) {
-    case 8:
-      if (type.code == kDLInt)
-        return static_cast<void*>(dst->mutable_data<int8_t>(dst_place));
-      if (type.code == kDLUInt)
-        return static_cast<void*>(dst->mutable_data<uint8_t>(dst_place));
-      PADDLE_THROW(platform::errors::Unimplemented(
-          "DLDataType code <%d> is illegal when DLDataType.bits is <%d>.",
-          type.code,
-          type.bits));
-    case 16:
-      if (type.code == kDLInt)
-        return static_cast<void*>(dst->mutable_data<int16_t>(dst_place));
-      if (type.code == kDLFloat)
-        return static_cast<void*>(
-            dst->mutable_data<paddle::platform::float16>(dst_place));
-      if (type.code == kDLBfloat)
-        return static_cast<void*>(
-            dst->mutable_data<paddle::platform::bfloat16>(dst_place));
-      PADDLE_THROW(platform::errors::Unimplemented(
-          "DLDataType code <%d> is illegal when DLDataType.bits is <%d>.",
-          type.code,
-          type.bits));
-    case 32:
-      if (type.code == kDLInt)
-        return static_cast<void*>(dst->mutable_data<int32_t>(dst_place));
-      if (type.code == kDLFloat)
-        return static_cast<void*>(dst->mutable_data<float>(dst_place));
-      PADDLE_THROW(platform::errors::Unimplemented(
-          "DLDataType code <%d> is illegal when DLDataType.bits is <%d>.",
-          type.code,
-          type.bits));
-    case 64:
-      if (type.code == kDLInt)
-        return static_cast<void*>(dst->mutable_data<int64_t>(dst_place));
-      if (type.code == kDLFloat)
-        return static_cast<void*>(dst->mutable_data<double>(dst_place));
-      if (type.code == kDLComplex)
-        return static_cast<void*>(
-            dst->mutable_data<paddle::platform::complex<float>>(dst_place));
-      PADDLE_THROW(platform::errors::Unimplemented(
-          "DLDataType code <%d> is illegal when DLDataType.bits is <%d>.",
-          type.code,
-          type.bits));
-    case 128:
-      if (type.code == kDLComplex)
-        return static_cast<void*>(
-            dst->mutable_data<paddle::platform::complex<double>>(dst_place));
-      PADDLE_THROW(platform::errors::Unimplemented(
-          "DLDataType code <%d> is illegal when DLDataType.bits is <%d>.",
-          type.code,
-          type.bits));
-    default:
-      PADDLE_THROW(platform::errors::Unimplemented(
-          "Unsupported DLDataType.bits %d.", type.bits));
-  }
-}
-
-void TensorFromDLPack(const ::DLTensor& dl_tensor, phi::DenseTensor* dst) {
-  platform::CPUPlace dst_place = platform::CPUPlace();
-  platform::CPUPlace src_place = platform::CPUPlace();
-
-  std::vector<int64_t> vec;
-  std::copy(dl_tensor.shape,
-            dl_tensor.shape + dl_tensor.ndim,
-            std::back_inserter(vec));
-
-  framework::DDim vddim = phi::make_ddim(vec);
-
-  dst->Resize(vddim);
-  ::DLDataType type = dl_tensor.dtype;
-  void* dst_ptr = GetDstPtrByDLDataType(type, dst, dst_place);
-
-  auto src_ptr = static_cast<const void*>(dl_tensor.data);
-  auto size = phi::product(vddim) * type.bits / 8;
-
-  if (dl_tensor.device.device_type == kDLCPU) {
-    memory::Copy(dst_place, dst_ptr, src_place, src_ptr, size);
-  }
-#if defined(PADDLE_WITH_CUDA) || defined(PADDLE_WITH_HIP)
-  if (dl_tensor.device.device_type == kDLGPU) {
-    platform::CUDAPlace dst_place =
-        platform::CUDAPlace(dl_tensor.device.device_id);
-    platform::CUDAPlace src_place =
-        platform::CUDAPlace(dl_tensor.device.device_id);
-    dst_ptr = GetDstPtrByDLDataType(type, dst, dst_place);
-    auto* ctx = platform::DeviceContextPool::Instance().GetByPlace(dst_place);
-    memory::Copy(dst_place,
-                 dst_ptr,
-                 src_place,
-                 src_ptr,
-                 size,
-                 reinterpret_cast<const phi::GPUContext&>(*ctx).stream());
-  }
-#endif
-#ifdef PADDLE_WITH_XPU
-  PADDLE_THROW(platform::errors::Unimplemented("XPUPlace is not supported"));
-#endif
-}
-
-void TensorFromDLPack(const DLManagedTensor* src, phi::DenseTensor* dst) {
-  std::vector<int64_t> vec;
-  std::copy(src->dl_tensor.shape,
-            src->dl_tensor.shape + src->dl_tensor.ndim,
-            std::back_inserter(vec));
-
-  framework::DDim vddim = phi::make_ddim(vec);
-  dst->Resize(vddim);
-  ::DLDataType type = src->dl_tensor.dtype;
-
-  auto src_ptr = static_cast<const void*>(src->dl_tensor.data);
-  auto size = phi::product(vddim) * type.bits / 8;
-
-  if (src->dl_tensor.device.device_type == kDLCPU) {
-    platform::CPUPlace dst_place = platform::CPUPlace();
-    platform::CPUPlace src_place = platform::CPUPlace();
-    void* dst_ptr = GetDstPtrByDLDataType(type, dst, dst_place);
-    memory::Copy(dst_place, dst_ptr, src_place, src_ptr, size);
-  }
-#if defined(PADDLE_WITH_CUDA) || defined(PADDLE_WITH_HIP)
-  if (src->dl_tensor.device.device_type == kDLGPU) {
-    platform::CUDAPlace dst_place =
-        platform::CUDAPlace(src->dl_tensor.device.device_id);
-    platform::CUDAPlace src_place =
-        platform::CUDAPlace(src->dl_tensor.device.device_id);
-    void* dst_ptr = GetDstPtrByDLDataType(type, dst, dst_place);
-    auto* ctx = platform::DeviceContextPool::Instance().GetByPlace(dst_place);
-    // Fix copy by share allocation.
-    memory::Copy(dst_place,
-                 dst_ptr,
-                 src_place,
-                 src_ptr,
-                 size,
-                 reinterpret_cast<const phi::GPUContext&>(*ctx).stream());
-  }
-#endif
-  src->deleter(const_cast<DLManagedTensor*>(src));
-#ifdef PADDLE_WITH_XPU
-  PADDLE_THROW(platform::errors::Unimplemented("XPUPlace is not supported"));
-#endif
-}
-
-template <typename T>
-std::string format_tensor(const phi::DenseTensor& tensor) {
-  // TODO(zhiqiu): use the print option to format tensor.
-  return "NOT IMPLEMENTED";
-}
-
-template <typename T>
-std::ostream& print_tensor(std::ostream& os, const phi::DenseTensor& tensor) {
-  auto inspect = tensor.data<T>();
-  auto element_num = tensor.numel();
-
-  os << "  - data: [";
-  // Note: int8_t && uint8_t is typedf of char, ostream unable to print properly
-  if (typeid(int8_t) == typeid(T) || typeid(uint8_t) == typeid(T)) {
-    if (element_num > 0) {
-      os << signed(inspect[0]);
-      for (int j = 1; j < element_num; ++j) {
-        os << " " << signed(inspect[j]);
-      }
-    }
-  } else {
-    if (element_num > 0) {
-      os << inspect[0];
-      for (int j = 1; j < element_num; ++j) {
-        os << " " << inspect[j];
-      }
-    }
-  }
-  os << "]";
-  return os;
-}
-
-template <>
-std::ostream& print_tensor<paddle::platform::complex<float>>(
-    std::ostream& os, const phi::DenseTensor& tensor) {
-  auto inspect = tensor.data<paddle::platform::complex<float>>();
-  auto element_num = tensor.numel();
-
-  os << "  - data: [";
-  if (element_num > 0) {
-    os << signed(inspect[0].real) << "+" << signed(inspect[0].imag) << "j";
-    for (int j = 1; j < element_num; ++j) {
-      os << " " << signed(inspect[j].real) << "+" << signed(inspect[j].imag)
-         << "j";
-    }
-  }
-  os << "]";
-  return os;
-}
-
-template <>
-std::ostream& print_tensor<paddle::platform::complex<double>>(
-    std::ostream& os, const phi::DenseTensor& tensor) {
-  auto inspect = tensor.data<paddle::platform::complex<double>>();
-  auto element_num = tensor.numel();
-
-  os << "  - data: [";
-  if (element_num > 0) {
-    os << signed(inspect[0].real) << "+" << signed(inspect[0].imag) << "j";
-    for (int j = 1; j < element_num; ++j) {
-      os << " " << signed(inspect[j].real) << "+" << signed(inspect[j].imag)
-         << "j";
-    }
-  }
-  os << "]";
-  return os;
-}
-
-std::ostream& operator<<(std::ostream& os, const LoD& lod) {
-  // NOTE(xiongkun):
-  // https://stackoverflow.com/questions/5195512/namespaces-and-operator-resolution
-  // if we don't redefine, the operator << of phi / framework LoD is not found.
-  paddle::string::operator<<(os, lod);
-  return os;
-}
 
 }  // namespace framework
 }  // namespace paddle
 
-namespace phi {
-
-std::ostream& operator<<(std::ostream& os, const LoD& lod) {
-  paddle::string::operator<<(os, lod);
-  return os;
-}
-
-std::ostream& operator<<(std::ostream& os, const phi::DenseTensor& t) {
-  if (t.lod().size() > 0) {
-    os << "  - lod: " << t.lod() << "\n";
-  }
-
-  os << "  - place: " << t.place() << "\n";
-  os << "  - shape: [" << t.dims() << "]\n";
-  os << "  - layout: " << phi::DataLayoutToString(t.layout()) << "\n";
-
-  DenseTensor tensor;
-  tensor.Resize(t.dims());
-  if (paddle::platform::is_cpu_place(t.place())) {
-    tensor.ShareDataWith(t);
-  } else {
-    paddle::platform::CPUPlace place;
-    paddle::framework::TensorCopy(t, place, &tensor);
-    paddle::platform::DeviceContextPool& pool =
-        paddle::platform::DeviceContextPool::Instance();
-    auto& dev_ctx = *pool.Get(t.place());
-    dev_ctx.Wait();
-  }
-
-#define PrintTensorCallback(cpp_type, proto_type)                 \
-  do {                                                            \
-    if (paddle::framework::TransToProtoVarType(tensor.dtype()) == \
-        proto_type) {                                             \
-      os << "  - dtype: " << proto_type << "\n";                  \
-      paddle::framework::print_tensor<cpp_type>(os, tensor);      \
-      return os;                                                  \
-    }                                                             \
-  } while (0)
-
-  _ForEachDataType_(PrintTensorCallback);
-  VLOG(1) << "PrintVar: unrecognized data type:" << t.type();
-  return os;
-}
-}  // namespace phi
