+++ paddle/fluid/operators/math/beam_search_npu.cc
@@ -572,7 +572,7 @@ class BeamSearchFunctor<platform::NPUDeviceContext, T> {
       PADDLE_THROW(platform::errors::InvalidArgument(
           "lod %s is not right in"
           " beam_search, please check your code.",
-          framework::LoDToString(lod)));
+          phi::LoDToString(lod)));
     }
     selected_ids->set_lod(lod);
     selected_scores->set_lod(lod);
